Projetos/Phishing-Domain-Detector/requirements.txt
requests
beautifulsoup4
tldextract
python-whois
fuzzywuzzy